+++ src/Utility/AssetsCreator.php
@@ -96,7 +96,7 @@ class AssetsCreator
             }
 
             $path = str_starts_with($path, '/') ? substr($path, 1) : $this->type . DS . $path;
-            $path = DS === '/' ? $path : $path = str_replace('/', DS, $path);
+            $path = DS === '/' ? $path : str_replace('/', DS, $path);
             $path = empty($plugin) ? WWW_ROOT . $path : Plugin::path($plugin) . 'webroot' . DS . $path;
 
             //Appends the file extension, if not already present