+++ tests/TestCase/Utility/AssetsCreatorTest.php
@@ -45,7 +45,7 @@ class AssetsCreatorTest extends TestCase
 
         $asset = $this->getProperty($asset, 'asset');
         $this->assertEquals(Configure::read('Assets.target'), dirname($asset));
-        $this->assertMatchesRegularExpression('/^[\d\w]+\.css$/', basename($asset));
+        $this->assertMatchesRegularExpression('/^\w+\.css$/', basename($asset));
 
         //With unsupported type
         $this->expectException(InvalidArgumentException::class);
@@ -157,7 +157,7 @@ class AssetsCreatorTest extends TestCase
     public function testCreateWithCss(): void
     {
         $result = (new AssetsCreator('test', 'css'))->create();
-        $this->assertMatchesRegularExpression('/^[\w\d]+$/', $result);
+        $this->assertMatchesRegularExpression('/^\w+$/', $result);
 
         $file = Configure::read('Assets.target') . DS . sprintf('%s.%s', $result, 'css');
         $this->assertFileExists($file);
@@ -167,7 +167,7 @@ class AssetsCreatorTest extends TestCase
 
         //Tests array
         $result = (new AssetsCreator(['test', 'test2'], 'css'))->create();
-        $this->assertMatchesRegularExpression('/^[\w\d]+$/', $result);
+        $this->assertMatchesRegularExpression('/^\w+$/', $result);
 
         $file = Configure::read('Assets.target') . DS . sprintf('%s.%s', $result, 'css');
         $this->assertFileExists($file);
@@ -192,7 +192,7 @@ class AssetsCreatorTest extends TestCase
     public function testCreateWithJs(): void
     {
         $result = (new AssetsCreator('test', 'js'))->create();
-        $this->assertMatchesRegularExpression('/^[\w\d]+$/', $result);
+        $this->assertMatchesRegularExpression('/^\w+$/', $result);
 
         $expected = 'function otherAlert(){alert("Another alert")}' . PHP_EOL .
             '$(()=>{const msg="Ehi!";alert(msg)})';
@@ -201,7 +201,7 @@ class AssetsCreatorTest extends TestCase
 
         //Tests array
         $result = (new AssetsCreator(['test', 'test2'], 'js'))->create();
-        $this->assertMatchesRegularExpression('/^[\w\d]+$/', $result);
+        $this->assertMatchesRegularExpression('/^\w+$/', $result);
 
         $expected = 'function otherAlert(){alert("Another alert")}' . PHP_EOL .
             '$(()=>{const msg="Ehi!";alert(msg)});' .
@@ -213,7 +213,7 @@ class AssetsCreatorTest extends TestCase
     }
 
     /**
-     * Test for `create()` method. The asset is created only if does not exist
+     * Test for `create()` method. The asset is created only if not exist
      * @uses \Assets\Utility\AssetsCreator::create()
      * @test
      */
@@ -227,7 +227,7 @@ class AssetsCreatorTest extends TestCase
         $time = filemtime($file);
 
         //Tries to create again the same asset. Now the creation time is the same
-        $result = (new AssetsCreator('test', 'css'))->create();
+        (new AssetsCreator('test', 'css'))->create();
         $this->assertEquals($time, filemtime($file));
 
         //Deletes asset and wait 1 second
@@ -235,7 +235,7 @@ class AssetsCreatorTest extends TestCase
         sleep(1);
 
         //Tries to create again the same asset. Now the creation time is different
-        $result = (new AssetsCreator('test', 'css'))->create();
+        (new AssetsCreator('test', 'css'))->create();
         $this->assertNotEquals($time, filemtime($file));
     }